test: benchmark perf if feature branches run everything locally
